refactor librarys

+++ neural_regress/SRP_torch_lib.py
@@ -197,7 +197,7 @@ def SparseRandomProjection_fit_transform_torch(X, n_components="auto", eps=0.1,
                                                     random_state=random_state, 
                                                     device=device, use_multinomial=use_multinomial)
     # transform the data, using sparse matrix multiplication with torch
-    featmat = torch.sparse.mm(X.to(device), projection_matrix.T).to("cpu")
+    featmat = torch.sparse.mm(X.float().to(device), projection_matrix.T).to("cpu") # note some x are half precision
     srp_sklearn = torch_to_sklearn_projection(projection_matrix, n_components)
     return featmat, srp_sklearn
